# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meem Market API",
        "description": "API documentation for Meem Market",
        "contact": {
            "email": "admin@meem-market.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        },
        {
            "url": "http://api.meem-market.com/",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/settings/{section}": {
            "get": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Get settings for a section",
                "description": "Returns all options for the specified section as key-value pairs.",
                "operationId": "getSettings",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "general",
                                "writing",
                                "reading",
                                "discussion",
                                "media",
                                "permalinks",
                                "privacy"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settings object"
                    },
                    "404": {
                        "description": "Unknown section"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Update settings for a section",
                "description": "Updates one or more options within the specified section.",
                "operationId": "updateSettings",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "general",
                                "writing",
                                "reading",
                                "discussion",
                                "media",
                                "permalinks",
                                "privacy"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Key-value pairs from the section's option list",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated"
                    },
                    "404": {
                        "description": "Unknown section"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token returned from /api/v1/admin/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}